Rename typo Allow specified type for given validation
